+++ bare-metal/fedora-coreos/kubernetes/versions.tf
@@ -8,7 +8,7 @@ terraform {
 
     ct = {
       source  = "poseidon/ct"
-      version = "~> 0.8"
+      version = "~> 0.9"
     }
 
     matchbox = {