+++ google-cloud/flatcar-linux/kubernetes/versions.tf
@@ -9,7 +9,7 @@ terraform {
 
     ct = {
       source  = "poseidon/ct"
-      version = "~> 0.8"
+      version = "~> 0.9"
     }
   }
 }